chore: move offline-player to server
