+++ permissions/build.gradle.kts
@@ -18,7 +18,7 @@ repositories {
 }
 
 dependencies {
-    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3")
+    compileOnly("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3")
     compileOnly(project(":server"))
     compileOnly(project(":sdk"))
     compileOnly("net.minestom:minestom-snapshots:$minestomVersion")
