+++ punishments/src/main/kotlin/gg/airbrush/punishments/commands/KickCommand.kt
@@ -12,9 +12,10 @@
 
 package gg.airbrush.punishments.commands
 
-import gg.airbrush.punishments.arguments.OfflinePlayerArgument
+import gg.airbrush.server.arguments.OfflinePlayerArgument
 import gg.airbrush.punishments.enums.PunishmentTypes
 import gg.airbrush.punishments.lib.*
+import gg.airbrush.server.arguments.OfflinePlayer
 import gg.airbrush.server.lib.mm
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.runBlocking
