+++ punishments/src/main/kotlin/gg/airbrush/punishments/commands/PunishmentsCommand.kt
@@ -13,9 +13,8 @@
 package gg.airbrush.punishments.commands
 
 import gg.airbrush.punishments.Punishment
-import gg.airbrush.punishments.arguments.OfflinePlayerArgument
+import gg.airbrush.server.arguments.OfflinePlayerArgument
 import gg.airbrush.punishments.enums.PunishmentTypes
-import gg.airbrush.punishments.lib.OfflinePlayer
 import gg.airbrush.punishments.punishmentConfig
 import gg.airbrush.sdk.SDK
 import gg.airbrush.sdk.classes.punishments.AirbrushPunishment
@@ -23,6 +22,7 @@ import gg.airbrush.sdk.lib.Placeholder
 import gg.airbrush.sdk.lib.Translations
 import gg.airbrush.sdk.lib.parsePlaceholders
 import gg.airbrush.sdk.lib.replaceTabs
+import gg.airbrush.server.arguments.OfflinePlayer
 import gg.airbrush.server.lib.mm
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.runBlocking
