+++ server/src/main/kotlin/gg/airbrush/server/arguments/OfflinePlayer.kt
@@ -10,7 +10,7 @@
  * You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package gg.airbrush.punishments.lib
+package gg.airbrush.server.arguments
 
 import java.util.UUID
 data class OfflinePlayer(val uniqueId: UUID, val username: String)
